# src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: DocuChat API
  version: 1.0.0
  description: AI-powered document Q&A system
...
paths:
...
  /documents/{id}:
    get:
      tags: [Documents]
      summary: Get a document
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Document
        "404":
          description: Not found
    delete:
      tags: [Documents]
      summary: Delete a document
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
...